+++ HeliumVotePlugin/components/VotingPowerCard.tsx
@@ -63,7 +63,6 @@ export const VotingPowerCard: React.FC<{
               <TokenDeposit
                 mint={councilMint}
                 tokenRole={GoverningTokenRole.Council}
-                councilVote={true}
                 setHasGovPower={setHasGovPower}
               />
             </div>
