+++ VoteStakeRegistry/components/Account/LockTokensAccount.tsx
@@ -415,7 +415,6 @@ const LockTokensAccount: React.FC<{
           <TokenDeposit
             mint={councilMint}
             tokenRole={GoverningTokenRole.Council}
-            councilVote={true}
             inAccountDetails={true}
           />
         </div>
